Derive knowledge from GitHub or GitLab repositories with the use of AI/LLM

Scan multiple Git repositories, pull specified files content and process it with Large Language Models. You can summarize the content in specific way, extract information and data, or find answers to your questions about the repositories. The output can be stored in vector database and used for semantic search or as a part of a RAG (Retrieval Augmented Generation) prompt.

Author

Maintainer: Kamil Wais kamil.wais@gmail.com

Authors:

Krystian Igras krystian8207@gmail.com

Maciej Banas banasmaciek@gmail.com